+++ server/routes/yelp.js
@@ -16,8 +16,7 @@ var request_yelp = function(set_parameters, callback) {
   var url = 'http://api.yelp.com/v2/search';
 
   var default_parameters = {
-    location: 'San+Francisco',
-    sort: '0'
+    location: 'San+Francisco'
   };
 
   var required_parameters = {
@@ -52,26 +51,7 @@ module.exports = function(db, passport, isLoggedIn) {
 
   var router = express.Router();
 
-  // =====================================
-  // FACEBOOK ROUTES =====================
-  // =====================================
-  // route for facebook authentication and login
-  router.get('/facebook', passport.authenticate('facebook'));
-
-  // handle the callback after facebook has authenticated the user
-  router.get('/facebook/callback',
-      passport.authenticate('facebook', {
-          successRedirect : '/profile',
-          failureRedirect : '/'
-  }));
-
-  // route for logging out
-  router.get('/logout', function(req, res) {
-    req.logout();
-    res.redirect('/');
-  });
-
-  router.get('/yelp', function(req, res) {
+router.get('/', function(req, res) {
     // set Yelp Search API params
     var params = {
       term: req.query.term,
@@ -90,15 +70,3 @@ module.exports = function(db, passport, isLoggedIn) {
   return router;
 
 };
-
-
-// route middleware to make sure a user is logged in
-function isLoggedIn(req, res, next) { // this is the only route that is middleware, and therefore needs to use the next parameter. 
-  // "next" will be the last function in a router.get/post etc request, and isLoggedIn will be added as a middle param
-  // e.g. router.get('/accounts', isLoggedIn, function(req, res) {...});
-  // if user is authenticated in the session, carry on
-  if (req.isAuthenticated())
-    return next();
-  // if they aren't redirect them to the home page
-  res.redirect('/');
-}
